back/requirements.txt
supabase
transformers
sentence-transformers
flask
flask-cors
PyMuPDF
ngrok
google-genai
python-dotenv